github-issues-integration-notion_src/requirements.txt
PyGithub
notion-client
